worker/requirements.txt
boto3==1.34.94
ffmpeg_python==0.2.0
librosa==0.10.1
Pillow==10.3.0
pysrt==1.1.2
redis==4.1.1
regex==2024.4.28
rq==1.10.0
Unidecode==1.3.8
webvtt_py==0.4.6
numpy<2.0.0,>=1.21.0